Typos fixes in Dashboard API doc
As Amplitude user I found typos in the documentation that I have been working with. The existing chart request won't work without the '/query' part after the 'chart_id', which is missing in the example. And the users request example has 'users?users?' part which causes the error, and should instead be 'users?' just one time.